requirements.txt
clickhouse-connect>=0.7.19
psycopg2-binary>=2.9.0
python-dateutil>=2.9.0
tqdm>=4.66.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
